[refactor] 유저 삭제 에러 핸들링 리팩토링
