+++ backend/src/main/java/com/coffee/backend/domain/auth/exception/InvalidTokenException.java
@@ -1,12 +1,8 @@
-package com.coffee.backend.exception;
+package com.coffee.backend.domain.auth.exception;
 
 public class InvalidTokenException extends RuntimeException {
 
     public InvalidTokenException(String message) {
         super(message);
-    }
-
-    public InvalidTokenException() {
-        super("유효하지 않은 토큰입니다");
     }
 }
