+++ backend/src/main/java/com/coffee/backend/domain/auth/exception/LoginIdAlreadyExistsException.java
@@ -1,4 +1,4 @@
-package com.coffee.backend.exception;
+package com.coffee.backend.domain.auth.exception;
 
 public class LoginIdAlreadyExistsException extends RuntimeException {
     public LoginIdAlreadyExistsException(String message) {
