+++ backend/src/main/java/com/coffee/backend/exception/CustomException.java
@@ -1,13 +1,10 @@
 package com.coffee.backend.exception;
 
+import com.coffee.backend.domain.auth.exception.ErrorCode;
+
 public class CustomException extends RuntimeException {
 
     private final ErrorCode code;
-
-    public CustomException(ErrorCode code, String message) {
-        super(code.getMessage() + "(" + message + ")");
-        this.code = code;
-    }
 
     public CustomException(ErrorCode code) {
         super(code.getMessage());
