--- backend/src/main/java/com/coffee/backend/exception/EmptyAuthorizationHeaderException.java
@@ -1,7 +1,0 @@
-package com.coffee.backend.exception;
-
-public class EmptyAuthorizationHeaderException extends RuntimeException {
-    public EmptyAuthorizationHeaderException() {
-        super("Authorization Header가 비어있습니다");
-    }
-}
